refactor: Move BB search function
